从redis中取出来的项目总预算和日预算转成double类型

+++ src/main/java/com/pxene/pap/service/CampaignService.java
@@ -496,7 +496,8 @@ public class CampaignService extends BaseService {
 					redisHelper.set(budgetKey, (int)newBudget);
 				} else {
 					if (redisHelper.exists(budgetKey)) {
-						int redisBudget = redisHelper.getInt(budgetKey) / 100;
+						// TODO : zytosee
+						double redisBudget = redisHelper.getInt(budgetKey) / 100;
 						int difBudget = newBudget - oldBudget;
 						if (difBudget < 0 && Math.abs(difBudget) > redisBudget) {
 							throw new IllegalArgumentException(PhrasesConstant.DIF_DAILY_BIGGER_REDIS);
